changed float to double in default tensor type for 3dcvae

+++ 3dcvae_pytotrch.py
@@ -11,7 +11,7 @@ import torch
 import torchmed
 import torch.utils.data
 from torch import nn, optim
-torch.set_default_tensor_type(torch.cuda.FloatTensor)  #comment out for cpu
+torch.set_default_tensor_type(torch.cuda.DoubleTensor)  #comment out for cpu
 from glob import glob
 from torch.autograd import Variable
 from torch.nn import functional as F
@@ -99,9 +99,6 @@ class VAE(nn.Module):
         
         self.mu = nn.Linear(h_dim, z_dim)
         self.logvar = nn.Linear(h_dim, z_dim)
-        
-        #classifier layer
-        #self.classify = nn.Linear(h_dim, n_classes)
 
         #decoder layers        
         self.linear = nn.Linear(z_dim, h_dim)
